Переносит отрисовку точек в отдельный контроллер
